fastapi-backend/app/requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
python-telegram-bot>=20.0
pydantic>=1.8.0
httpx>=0.24.0
python-dotenv>=0.19.0
